% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_functions.r
\name{all.in.one}
\alias{all.in.one}
\title{all.in.one}
\usage{
\method{all}{in.one}(UseSubset = "SAC", filename = "SAC.pdf",
  reportNR = TRUE)
}
\arguments{
\item{UseSubset}{Corresponds to a column name in outputplan holding flags (all.in.one)}

\item{filename}{common_root.pdf or common_root.csv}

\item{reportNR}{If TRUE, a plot with missing figure numbers and titles is produced}
}
\description{
Produces pdf files with batches of graphics based on flags in outputplan. A progress bar is displayed.
}
\details{
Prerequisites: You need to have output, code, data directory paths defined in your workspace. These should take variable names od, cd, dd, respectively. This can be done by running a personalized set of the following commands:

Code directory needs to hold the .r files associated with the subset of figures to be produced.

Suggest running outputplan.report() first. A progress bar also helps to see run is incomplete. A manual check on the total number of pages in the final pdf should be made.
}
\section{Value}{

\describe{
A .pdf file called filename.pdf is deposited in the output directory.
}
}
\author{
Greg Cicconetti
}

