% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{benrisk2.data}
\alias{benrisk2.data}
\title{This is a dataset structured for building figures using forest.plot}
\description{
This is a dataset structured for building figures using forest.plot
}
\author{
Greg Cicconetti
}
