\name{boxplot.driver}
\alias{boxplot.driver}
\docType{data}
\title{
boxplot.driver}
\description{
boxplot driver file 
}
\usage{data("boxplot.driver")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details
}

\examples{
data(boxplot.driver)
## maybe str(boxplot.driver) ; plot(boxplot.driver) ...
}
\keyword{datasets}
