% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{check.ggplot.outliers}
\alias{check.ggplot.outliers}
\title{check.ggplot.outliers}
\usage{
check.ggplot.outliers(plot.object = NULL)
}
\arguments{
\item{plot.object}{the ggplot object to check}
}
\description{
Reports via cat statements when ggplot windows truncate data
}
\details{
Used in conjunction with log files created with start_session_log
}
\author{
David Wade
}
