% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{code.column.names}
\alias{code.column.names}
\title{code column names}
\usage{
code.column.names(parent.df = working.df, format.name = "TRT633")
}
\arguments{
\item{parent.df}{The name of the dataframe to be processed.}

\item{format.name}{The name of a pre-specified format}
}
\description{
This function reads in a dataframe from a HARP dddata datafile .csv and returns a copy of the dataframe with a format-specified subset of the column names coded to a format-specified set of numerical values specified by the format name passed.
}
\author{
David Wade
}

