% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{code.treatment.values}
\alias{code.treatment.values}
\title{code.treatment.values}
\usage{
code.treatment.values(parent.df = working.df, format.name = "TRT633",
  coded.col = "TREATMENT", raw.col = "Planned.Treatment")
}
\arguments{
\item{format.name}{}

\item{coded.col}{}

\item{raw.col}{}
}
\description{
blah
}

