% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{convert.dates}
\alias{convert.dates}
\title{convert.dates}
\usage{
convert.dates(dataset)
}
\arguments{
\item{dataset}{a data.frame}
}
\description{
This function applies ddmmmyy.data to all columns of a data.frame ending in 'DT'
}
\examples{
\dontrun{
}
}
\author{
Greg Cicconetti
}

