% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{ddmmmyy.data}
\alias{ddmmmyy.data}
\title{ddmmyy.chron}
\usage{
ddmmmyy.data(string)
}
\arguments{
\item{string}{character string vector with ddmmyy format}
}
\description{
This function converts text dates to Date dates via lubridate
}
\examples{
\dontrun{
}
}
\author{
Greg Cicconetti
}
\seealso{
convert.dates
}

