% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{decode.column.values}
\alias{decode.column.values}
\title{decode.column.values}
\usage{
decode.column.values(parent.df = working.df, format.name = "TRT633",
  coded.col = "TREATMENT", formatted.col = "TREATMENT.TEXT",
  formatted.symbol.col = "TREATMENT.SYMBOL")
}
\arguments{
\item{parent.df}{The name of the dataframe to be processed}

\item{format.name}{The name of a pre-specified treatment format}

\item{coded.col}{The name of the existing column containing the numerically coded treatment values}

\item{formatted.col}{The name of the new column to contain the decoded treatment text values}

\item{formatted.symbol.col}{The name of the new column to contain the decoded treatment plot symbols}
}
\description{
This function reads in a dataframe that has a column with numerically coded values and returns a copy of the dataframe with a new column containing the decoded text values and the decoded plot symbols for those same values
}
\examples{
\dontrun{
# Assign formatted values to TREATMENT
working.df <- decode.column.values (parent.df = working.df, format.name  ="TRT633", coded.col ="TREATMENT", formatted.col="TREATMENT.TEXT", formatted.symbol.col="TREATMENT.SYMBOL")
working.df <- (parent.df=working.df, format.name="TRT633")
}
}
\author{
David Wade
}

