% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{decode.treatment.values}
\alias{decode.treatment.values}
\title{decode.treatment.values}
\usage{
decode.treatment.values(parent.df = working.df, format.name = "TRT633",
  coded.col = "TREATMENT", formatted.col = "TREATMENT.TEXT",
  formatted.symbol.col = "TREATMENT.SYMBOL")
}
\arguments{
\item{parent.df}{The name of the dataframe to be processed}

\item{format.name}{The name of a pre-specified treatment format (valid values are: ?TRT633?)}
}
\description{
This function reads in a dataframe that has a column with numerically coded values and returns a copy of the dataframe with a new column containing the decoded text values and the decoded plot symbols for those same values
}
\examples{
\dontrun{    }
}
\author{
David Wade
}

