% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{default.settings}
\alias{default.settings}
\title{default.settings}
\usage{
default.settings(
  pos = 1,
  envir = as.environment(pos),
  my.path = getwd(),
  main.theme = "theme_bw",
  page.width = 11,
  page.height = 8.5,
  right.margin = 0.75,
  left.margin = 0.75,
  top.margin = 1.4 - 0.5,
  bottom.margin = 1.75 - 0.5
)
}
\arguments{
\item{pos}{used internally by some functions}

\item{envir}{used internally by some functions}

\item{my.path}{path to main directory,}

\item{main.theme}{text string name of theme to be called by theme_set,}

\item{page.width}{used by build.page and annotate.page; presumed to be inches}

\item{page.height}{used by build.page and annotate.page; presumed to be inches}

\item{right.margin}{used by build.page and annotate.page; presumed to be inches}

\item{left.margin}{used by build.page and annotate.page; presumed to be inches}

\item{top.margin}{used by build.page and annotate.page; presumed to be inches}

\item{bottom.margin}{used by build.page and annotate.page}
}
\value{
This function assigns character string objects to the global environment.
}
\description{
Global Defaults
}
\details{
Global Defaults
}
\section{Value}{
 
\describe{
The following are assigned to global environment upon calling:
 \itemize{
 \item my.path
 \item dd
 \item cd
 \item od
 \item blankPanel 
 \item page.width 
 \item page.height 
 \item right.margin 
 \item left.margin 
 \item top.margin 
 \item bottom.margin 
 \item graph.region.h
 \item graph.region.w
}
}
}

\author{
Greg Cicconetti
}
