% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global.r
\name{default.settings2}
\alias{default.settings2}
\title{default.settings2 - next iteration of default.settings}
\usage{
default.settings2(my.path = getwd(), main.theme = "theme_bw",
  dd = paste0(getwd(), "/dddata/"), cd = paste0(getwd(), "/code/"),
  od = paste0(getwd(), "/output/"), logd = paste0(getwd(), "/log/"),
  page.width = 11, page.height = 8.5, right.margin = 0.75,
  left.margin = 0.75, top.margin = 1.4 - 0.5, bottom.margin = 1.75 - 0.5)
}
\arguments{
\item{my.path}{path to main directory,}

\item{main.theme}{text string name of theme to be called by theme_set,}

\item{dd}{directory where data is stored}

\item{cd}{directory where driver (code) files are stored}

\item{od}{directory where output files are sent}

\item{logd}{directory where log files are sent}

\item{page.width}{presumed to be inches}

\item{page.height}{presumed to be inches}

\item{right.margin}{presumed to be inches}

\item{left.margin}{presumed to be inches}

\item{top.margin}{presumed to be inches}

\item{bottom.margin}{presumed to be inches}
}
\description{
Global Defaults
}
\details{
Global Defaults
}
\section{Value}{

\describe{
The following are assigned to global environment upon calling:
 \itemize{
 my.path <<- getwd()
  \item dd
  \item cd
  \item od
  \item blankPanel
  \item page.width
  \item page.height
  \item right.margin
  \item left.margin
  \item top.margin
  \item bottom.margin
  \item graph.region.h
  \item graph.region.w
  \item blankPanel #'
}
}
}
\author{
Greg Cicconetti
}

