% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{demog.data}
\alias{demog.data}
\title{This is a dataset structured for building figures using bar.plot, box.plot, and cdf.plot}
\description{
This is a dataset structured for building figures using bar.plot, box.plot, and cdf.plot
}
\author{
Greg Cicconetti
}
