\name{driver1}
\alias{driver1}
\docType{data}
\title{
driver1}
\description{
driver1
}
\usage{data("driver1")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver1)
## maybe str(driver1) ; plot(driver1) ...
}
\keyword{datasets}
