\name{driver10}
\alias{driver10}
\docType{data}
\title{
driver10}
\description{
driver10
}
\usage{data("driver10")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver10)
## maybe str(driver10) ; plot(driver10) ...
}
\keyword{datasets}
