\name{driver3}
\alias{driver3}
\docType{data}
\title{
driver3}
\description{
driver3
}
\usage{data("driver3")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver3)
## maybe str(driver3) ; plot(driver3) ...
}
\keyword{datasets}
