\name{driver4}
\alias{driver4}
\docType{data}
\title{
driver4}
\description{
driver4
}
\usage{data("driver4")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver4)
## maybe str(driver4) ; plot(driver4) ...
}
\keyword{datasets}
