\name{driver5}
\alias{driver5}
\docType{data}
\title{
driver5}
\description{
driver5
}
\usage{data("driver5")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver5)
## maybe str(driver5) ; plot(driver5) ...
}
\keyword{datasets}
