\name{driver7}
\alias{driver7}
\docType{data}
\title{
driver7}
\description{
driver7
}
\usage{data("driver7")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver7)
## maybe str(driver7) ; plot(driver7) ...
}
\keyword{datasets}
