\name{driver8}
\alias{driver8}
\docType{data}
\title{
driver8}
\description{
driver8
}
\usage{data("driver8")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver8)
## maybe str(driver8) ; plot(driver8) ...
}
\keyword{datasets}
