% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{fmt}
\alias{fmt}
\title{fmt}
\usage{
fmt(digits = 2)
}
\arguments{
\item{digits}{number of digits displayed}
}
\description{
A function to control number of digits used in graphics.
}
\details{
This function is used within ggplot, e.g. (scale_y_continuous(labels=fmt(digits=3))) to control the number of digits presented. By default, axis labels will truncate zeros so that labels might read: 0, 2.5, 5, 7.5. Using this will result in labels: 0.0, 2.5, 5.0, 7.5.
}
\author{
Greg Cicconetti
}
