% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{funnel.plot}
\alias{funnel.plot}
\title{funnel.plot}
\usage{
funnel.plot(parent.df = read.csv(paste(dd, "g_KMfmace.csv", sep = "")),
  censored.time = "centime", censor.var = "censor", base = "region",
  reference = "global", region = "Asia/Pacific", country = "Hong Kong",
  voffset = 0, text.size = 4)
}
\arguments{
\item{base}{Could be "centre", "country", or "region"}

\item{reference}{should be at least one level above base:  "country", "region", "global"}

\item{region}{if reference == "region", specify}

\item{country}{if reference == "country", specify}

\item{voffset}{vertical offset for text}

\item{response}{Column holding the count variable (in case of 1st MACE: 0 censored, 1 observed)}
}
\description{
Creates a funnel plot tailored for regions.
}
\details{
Appeals to Poisson-distribution for confidence bounds and makes continuity correction. Outliers are identified by name.  # This function should return a series of ggplot objects
The number of graphics returned should depend on the reference.category: If reference category is "global" one graphic is returned; number of points equals number of unique categories.
If reference category is region, then one graphic is returned for each region.  When category is country, a dot for each country is used. When category is centreid a dot for each centreid is used.
If reference category is country, then one graphic for each country is returned.  Category is assumed to be centreid; one dot for each centre.
}
\section{Value}{

\describe{
A revamped version of the older funnel plot.  Previous version produced funnel plots for regions relative to  study's global average.  This version allows comparison of 'base' relative to a 'reference'.
}
}
\author{
Greg Cicconetti
}
\references{
David J. Spiegelhalter. Funnel plots for comparing institutional performance. Statist. Med. 2005; 24:1185-1202
}

