% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{insertNL}
\alias{insertNL}
\title{Insert a New Line}
\usage{

  insertNL(text.string = "Kaplan-Meier Plot of Time to First Occurence of the Total Coronary Events Excluding Target Lesion Revascularizations in Subjects Treated with Percutaneous Coronary Intervention Prior to Randomization During the Time Period for Follow-up of Cardiovascular Efficacy Endpoints",
  nlhere = seq(7, 70, 7))
}
\arguments{
\item{text.string}{a text string holding the title to a display}

\item{nlhere}{a vector of integers identifying after which ordered word in the title a new line should be placed}
}
\description{
This function takes a string and returns a list containing 4 components.
}
\details{
Fill in details
}
\examples{
\dontrun{
insertNL(outputplan$FigureTitle[90])
}
}
\author{
Greg Cicconetti
}

