% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{jitter.plot}
\alias{jitter.plot}
\title{jitter.plot - REVISIT - should sync with box.plot}
\usage{
jitter.plot(parent.df = working.df, y.col = "VSBMI",
  y.label = "BMI (m/kg^2)", category.col = "TRTGRP",
  category.label = "Treatment Group", y.limits = c(10, 100),
  y.ticks = seq(10, 100, 10), y.digits = 0, shape.palette = bplot.shapes,
  category.palette = treatment.palette)
}
\arguments{
\item{y.col}{parent.df column associated with response vairable}

\item{y.label}{value gets passed to labs}

\item{category.col}{parent.df column associated with explantory variable}

\item{category.label}{passed to x-axis label}

\item{y.limits}{passed to scale_y_continuous}

\item{y.ticks}{passed to scale_y_continuous}

\item{y.digits}{passed to scale_y_continuous label's, fmt}

\item{shape.palette}{values passed to scale_shape_manual}

\item{category.palette}{colors assoicated with categorical variable}
}
\description{
Produces boxplots assuming TRTGRP is the explantory variable.
}
\details{
Adjust y.limits, y.ticks and y.digits at the the driver level. Note: This function computes summary statistics from raw data.
}
\author{
Greg Cicconetti
}

