% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{km.data}
\alias{km.data}
\title{This is a dataset structured for building figures using km.plot}
\description{
This is a dataset structured for building figures using km.plot
}
\author{
Greg Cicconetti
}
