% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{lasso.plot}
\alias{lasso.plot}
\title{lasso.plot}
\usage{
lasso.plot(lasso.model = lasso.mod)
}
\arguments{
\item{lasso.model}{glmnet object}
}
\description{
This function takes a glmnet object and builds a data.frame create ggplot versions of glmnet plot methods - expand to offer 3 standards and return a list
}
\author{
Greg Cicconetti
}

