\name{lineplot.data}
\alias{lineplot.data}
\docType{data}
\title{
lineplot.data}
\description{
lineplot.data}
\usage{data("lineplot.data")}
\format{
  A data frame with 190 observations on the following 17 variables.
  \describe{
    \item{\code{Analysis.Visit..N.}}{a numeric vector}
    \item{\code{Analysis.Visit}}{a factor with levels \code{BASELINE} \code{DAY 1} \code{SCREEN WEEK -2} \code{SCREEN WEEK -4} \code{WEEK 10} \code{WEEK 12} \code{WEEK 14} \code{WEEK 16} \code{WEEK 18} \code{WEEK 2} \code{WEEK 20} \code{WEEK 21} \code{WEEK 22} \code{WEEK 23} \code{WEEK 24} \code{WEEK 28 FOLLOW-UP} \code{WEEK 4} \code{WEEK 6} \code{WEEK 8}}
    \item{\code{tt_segorder}}{a numeric vector}
    \item{\code{X.}}{a factor with levels \code{Analysis Value}}
    \item{\code{Order.of.Statistical.List}}{a numeric vector}
    \item{\code{X..1}}{a factor with levels \code{25th Percentile} \code{75th Percentile} \code{LCLM} \code{Max.} \code{Mean} \code{Median} \code{Min.} \code{n} \code{SD} \code{UCLM}}
    \item{\code{Summary.Level.Variable.Added.by.TU_STATSWITHTOTAL}}{a numeric vector}
    \item{\code{NAME.OF.FORMER.VARIABLE}}{a factor with levels \code{TT_RESULT}}
    \item{\code{LABEL.OF.FORMER.VARIABLE}}{a factor with levels \code{Result - formatted}}
    \item{\code{Not.Assigned..N.3.}}{a numeric vector}
    \item{\code{Control..N.10.}}{a numeric vector}
    \item{\code{X4.mg..N.11.}}{a numeric vector}
    \item{\code{X6.mg..N.16.}}{a numeric vector}
    \item{\code{X8.mg..N.8.}}{a numeric vector}
    \item{\code{X10.mg..N.11.}}{a numeric vector}
    \item{\code{X12.mg..N.6.}}{a numeric vector}
    \item{\code{X.N.1.}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(lineplot.data)
## maybe str(lineplot.data) ; plot(lineplot.data) ...
}
\keyword{datasets}
