% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{outputplan}
\alias{outputplan}
\title{This is a dataset structured to facilitate mass figure production}
\description{
This is a dataset structured to facilitate mass figure production
}
\author{
Greg Cicconetti
}
