% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global.r
\name{process.bslchar}
\alias{process.bslchar}
\title{process.bslchar}
\usage{
process.bslchar(dframe = read.csv(paste(dd, "g_bslchar.csv", sep = "")))
}
\arguments{
\item{dframe}{The input data.frame should be demog.data.}
}
\description{
This function processes the demog.data set included in this package. Serves as an example for preprocessing data prior to building graphics.
}
\section{Value}{

\describe{
A data.frame.
}
}
\examples{
\dontrun{
data(demog.data)
default.settings()
process.bslchar(dframe = demog.data)
}
}
\author{
Greg Cicconetti
}

