% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{refresh.outputplan}
\alias{refresh.outputplan}
\title{Refresh the Output Plan}
\usage{
refresh.outputplan(
  loadplan = TRUE,
  filename = "outputplan.csv",
  pos = 1,
  envir = as.environment(pos)
)
}
\arguments{
\item{loadplan}{logical; if TRUE then it loads from the filename}

\item{filename}{common_root.pdf or common_root.csv}

\item{pos}{used internally by some functions}

\item{envir}{used internally by some functions}
}
\value{
This function returns a data.frame.
}
\description{
Reloads outputplan_study.csv file and applies canonical formatting changes.
}
\details{
Ensure all columns are read in as character vectors. Ensure all missing entries are replaced with blank character string. Ensure all escape characters for carrige returns are respected. Grabs the 'modified time' from file attributes associated with .csv files named in the outputplan.
}
\author{
Greg Cicconetti
}
