% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{start_session_log}
\alias{start_session_log}
\title{start_session_log}
\usage{
start_session_log(
  x,
  outputfile = "example.PDF",
  pos = 1,
  envir = as.environment(pos),
  ...
)
}
\arguments{
\item{x}{used internally}

\item{outputfile}{passed to name the session history log file}

\item{pos}{used internally by some functions}

\item{envir}{used internally by some functions}

\item{...}{additional params}
}
\value{
This function works in conjunction with stop_session_log to create a log file.
}
\description{
A function to start logging the session history for a graphic driver run
}
\details{
Note that the stop_session_log function is used to stop the logging and save the log file.
}
\section{Value}{
 
\describe{
No objects are returned by this function.
}
}

\author{
David Wade
}
