% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{stop.session.log}
\alias{stop.session.log}
\title{stop.session.log}
\usage{
stop.session.log()
}
\description{
A function to stop logging the session history for a graphic driver run and save the session history file
}
\details{
Note that the start.session.log function is used to start the logging, and it must be called first.
}
\section{Value}{

\describe{
No objects are returned by this function.
}
}
\examples{
\dontrun{ }
}
\author{
David Wade
}

