% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{stop_session_log}
\alias{stop_session_log}
\title{stop_session_log}
\usage{
stop_session_log()
}
\value{
This function works in conjunction with start_session_log to create a log file.
}
\description{
A function to stop logging the session history for a graphic driver run and save the session history file
}
\details{
Note that the start_session_log function is used to start the logging, and it must be called first.
}
\section{Value}{
 
\describe{
No objects are returned by this function.
}
}

\author{
David Wade
}
