% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{summary.lineplot.data}
\alias{summary.lineplot.data}
\title{This is a dataset that would need some pre-processing ahead of using line.plot}
\description{
This is a dataset that would need some pre-processing ahead of using line.plot
}
\author{
Greg Cicconetti
}
