% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{sync.ylab.widths}
\alias{sync.ylab.widths}
\title{sync.ylab.widths}
\usage{
sync.ylab.widths(gg.list, default.length = 2)
}
\arguments{
\item{gg.list}{a list of ggplot objects (sync.ylab.widths)}

\item{default.length}{set to 2}
}
\value{
A ggplot object is returned.
}
\description{
Aligns the widths of ggplot objects to ensure common plot regions. The maximum length required for y-axis labels among the list is determined and applied to the other plots. This assists in syncing the widths of ggplot objects for the purpose of align figures on a page.
}
\author{
Greg Cicconetti
}
