% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.r
\name{theme_grey2_nomargins}
\alias{theme_grey2_nomargins}
\alias{theme_grey2_default_margins}
\alias{theme_bw2_nomargins}
\alias{theme_bw2_default_margins}
\alias{theme_table_nomargins}
\title{figuRes2 themes}
\usage{
theme_grey2_nomargins(base_size = 12, base_family = "")

theme_grey2_default_margins(base_size = 12, base_family = "")

theme_bw2_nomargins(base_size = 12, base_family = "")

theme_bw2_default_margins(base_size = 12, base_family = "")

theme_table_nomargins(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{used in set_theme calls}

\item{base_family}{used in set_theme calls}
}
\value{
The returns a function that can be passed to ggplot2::theme_set
}
\description{
Adapts theme_grey() found in ggplot2
}
\details{
axis.text colour changed from "grey50" to "black"; legend.position changed from "right" to "bottom"; legend.direction changed to "horizontal"; plot.margin changed from default unit(c(1, 1, 0.5, 0.5), "lines") to unit(c(0, 0, 0, 0), "in")
}
\section{Functions}{
\itemize{
\item \code{theme_grey2_default_margins()}: Same as theme_grey2_nomargins but with margins set to ggplot defaults, unit(c(1, 1, 0.5, 0.5), "lines")

\item \code{theme_bw2_nomargins()}: Similar to theme_grey2

\item \code{theme_bw2_default_margins()}: Similar to theme_bw_nomargins but with margins set to ggplot defaults, unit(c(1, 1, 0.5, 0.5), "lines")

\item \code{theme_table_nomargins()}: alteration to theme_grey

}}
\examples{
{
ggplot2::theme_set(theme_grey2_nomargins())
}
}
\author{
Greg Cicconetti
}
