% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{box.plot}
\alias{box.plot}
\title{box.plot}
\usage{
box.plot(parent.df, y.col = "VSBMI", y.label = "BMI (m/kg^2)",
  category.col = "TRTGRP", category.label = "Treatment Group",
  y.limits = c(10, 100), y.ticks = seq(10, 100, 10), y.digits = 0,
  shape.palette = c(1, 2), category.palette = c("red", "blue"),
  bplot.text = 3)
}
\arguments{
\item{parent.df}{data.frame used by ggplot}

\item{y.col}{parent.df column associated with response vairable}

\item{y.label}{value gets passed to labs}

\item{category.col}{parent.df column associated with explantory variable}

\item{category.label}{passed to x-axis label}

\item{y.limits}{passed to scale_y_continuous}

\item{y.ticks}{passed to scale_y_continuous}

\item{y.digits}{passed to scale_y_continuous label's, fmt}

\item{shape.palette}{values passed to scale_shape_manual}

\item{category.palette}{colors assoicated with categorical variable}
}
\description{
Produces boxplots
}
\details{
Adjust y.limits, y.ticks and y.digits at the the driver level. Note: This function computes summary statistics from raw data.
}
\examples{
\dontrun{
#' data(demog.data)
# pre-processing

levels(demog.data$SEX) <- c("Female", "Male")

# p1 <- box.plot(parent.df = demog.data,
#     y.col = "VSBMI",
#     y.label = expression(paste("BMI (m/kg",phantom()^2,")")),
#     category.col = "SEX",
#     category.label = "Gender",
#      y.limits = c(0, 70),
#      y.ticks = seq(0, 100, 10),
#      y.digits = 0,
#      shape.palette = c(20, 20),
#      category.palette = rainbow(6),
#      bplot.text = 4)
# print(p1)
}
}
\author{
Greg Cicconetti
}

