% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/builders.R
\name{build.page}
\alias{build.page}
\title{build.page}
\usage{
build.page(interior.h = c(1/3, 1/3, 1/3), interior.w = c(0.5, 0.25, 0.25),
  ncol = 3, nrow = 4, interior = list(synced.fps[[1]], dtest1, ttest1,
  synced.fps[[2]], dtest2, ttest2, synced.fps[[3]], dtest3, ttest3,
  synced.fps[[4]], dtest4, ttest4), test.dim = FALSE, page.height = 8.5,
  page.width = 11, right.margin = 0.75, left.margin = 0.75,
  top.margin = 1.4 - 0.5, bottom.margin = 1.75 - 0.5)
}
\arguments{
\item{interior.h}{a vector summing to 1 to indicate how to partition the heights}

\item{interior.w}{a vector summing to 1 to indicate how to partition the widths}

\item{ncol}{number of columns for the grid of graphics being built by build.page}

\item{nrow}{number of rows for the grid of graphics being built by build.page}

\item{interior}{a list of nrow*ncol grobs/ggplot objects to be displayed in the grid, ordered by row then col}

\item{test.dim}{logical. Assists with figure development. If TRUE it makes a call to grid.show.layout.}

\item{page.height}{used by build.page and annotate.page}

\item{page.width}{used by build.page and annotate.page}

\item{right.margin}{used by build.page and annotate.page}

\item{left.margin}{used by build.page and annotate.page}

\item{top.margin}{used by build.page and annotate.page}

\item{bottom.margin}{used by build.page and annotate.page}
}
\description{
Takes page dimensions, figure layout dimenesions and an ordered list of grobs/ggplot objects orients them on a page
}

