% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{cdf.data}
\alias{cdf.data}
\title{This is a dataset structured for building figures using cdf.plot}
\description{
This is a dataset structured for building figures using cdf.plot
}
\author{
Greg Cicconetti
}
\keyword{cdf.plot}
\keyword{dataset,}

