% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{cdf.plot}
\alias{cdf.plot}
\title{cdf.plot}
\usage{
cdf.plot(parent.df = working.df, category.col = "TRTGRP",
  category.label = "Treatment Group", response.col = "EXDURMTH",
  x.label = "Month 6 HDL Cholesterol", x.limits = c(0, 60),
  x.ticks = seq(0, 60, 10),
  y.label = "Percentage with 3 Month HDL Cholesterol Less Than x-value",
  y.limits = c(0, 1), y.ticks = seq(0, 1, 0.2), line.size = 0.75,
  category.palette = c("red", "blue"), exclude = c(0))
}
\arguments{
\item{category.col}{parent.df column associated with explantory variable}

\item{category.label}{passed to x-axis label}

\item{x.label}{value gets passed to labs}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{y.label}{value gets passed to labs}

\item{y.limits}{passed to scale_y_continuous}

\item{y.ticks}{passed to scale_y_continuous}

\item{line.size}{value gets passed to size within geom_line, geom_step}

\item{category.palette}{colors assoicated with categorical variable}

\item{exclude}{}
}
\description{
Function to produce Cumulative Distribution plot.  Statistics computed by stat_ecdf().
}
\examples{
\dontrun{
# data(demog.data)
# cdf.plot(parent.df= demog.data,
#         category.col = "SEX",
#         category.label   = "Gender",
#         response.col = "VSBMI",
#         x.label = expression(paste("BMI (m/kg",phantom()^2,")")),
#         x.limits=c(0,60),
#         x.ticks=seq(0,60,5),
#         y.label = "Percentage of Subjects",
#         y.limits= c(0,1),
#         y.ticks = seq(0,1,.2),
#         line.size=.75,
#         category.palette=c("red", "blue"),
#         exclude=c(0))
}
}
\author{
Greg Cicconetti
}

