% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_functions.r
\name{check.ggplot.outliers}
\alias{check.ggplot.outliers}
\title{check.ggplot.outliers}
\usage{
check.ggplot.outliers(plot.object = NULL)
}
\description{
Reports via cat statements when ggplot windows truncate data
}
\details{
Used in conjunction with log files created with start.session.log
}
\author{
David Wade
}

