% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_functions.r
\name{default.settings}
\alias{default.settings}
\title{default.settings}
\usage{
default.settings(pos = 1, envir = as.environment(pos))
}
\arguments{
\item{pos}{used internally by some functions}

\item{envir}{used internally by some functions}
}
\description{
Global Defaults
}
\details{
Global Defaults
}
\section{Value}{

\describe{
Describe
}
}
\author{
Greg Cicconetti
}

