% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/global.r
\name{default.settings}
\alias{default.settings}
\title{default.settings}
\usage{
default.settings()
}
\description{
Global Defaults
}
\details{
Global Defaults
}
\section{Value}{

\describe{
The following are assigned to global environment upon calling:
 \itemize{
 my.path <<- getwd()
  \item dd
  \item cd
  \item od
  \item blankPanel
}
}
}
\author{
Greg Cicconetti
}

