\name{driver2}
\alias{driver2}
\docType{data}
\title{
driver2}
\description{
driver2
}
\usage{data("driver2")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver2)
## maybe str(driver2) ; plot(driver2) ...
}
\keyword{datasets}
