\name{driver6}
\alias{driver6}
\docType{data}
\title{
driver6}
\description{
driver6
}
\usage{data("driver6")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver6)
## maybe str(driver6) ; plot(driver6) ...
}
\keyword{datasets}
