\name{driver9}
\alias{driver9}
\docType{data}
\title{
driver9}
\description{
driver9
}
\usage{data("driver9")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
No details.}

\examples{
data(driver9)
## maybe str(driver9) ; plot(driver9) ...
}
\keyword{datasets}
