% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{facetAdjust}
\alias{facetAdjust}
\title{FacetLabelAdjuster}
\usage{
facetAdjust(x, pos = c("up", "down"), newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{x}{a ggplot object}

\item{pos}{maintain default}

\item{newpage}{maintain default}

\item{vp}{maintain default}
}
\description{
This function takes a 'facet wrapped' ggplot and adds axis labels when a rxc grid is incomplete
}
\details{
Taken in whole from: http://stackoverflow.com/questions/13297155/add-floating-axis-labels-in-facet-wrap-plot
}
\examples{
\dontrun{
}
}
\author{
Ripped from stackoverflow
}

