% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all_functions.r
\docType{data}
\name{forest.data}
\alias{forest.data}
\title{This is a dataset structured for building figures using forest.plot}
\description{
This is a dataset structured for building figures using forest.plot

This is a dataset structured for building figures using km.plot
}
\author{
Greg Cicconetti

Greg Cicconetti
}
\keyword{dataset,}
\keyword{forest.plot}
\keyword{km.plot}

