\name{forest.data}
\alias{forest.data}
\docType{data}
\title{
forest.data}
\description{
A data set for producting forest plots. }
\usage{data("forest.data")}
\format{
  A data frame with 124 observations on the following 16 variables.
  \describe{
    \item{\code{skp}}{a numeric vector}
    \item{\code{skp1}}{a numeric vector}
    \item{\code{revord}}{a numeric vector}
    \item{\code{subgrpcd_order}}{a numeric vector}
    \item{\code{n1}}{a numeric vector}
    \item{\code{n2}}{a numeric vector}
    \item{\code{e1}}{a numeric vector}
    \item{\code{p1}}{a numeric vector}
    \item{\code{e2}}{a numeric vector}
    \item{\code{p2}}{a numeric vector}
    \item{\code{hr}}{a numeric vector}
    \item{\code{low}}{a numeric vector}
    \item{\code{high}}{a numeric vector}
    \item{\code{intchi}}{a numeric vector}
    \item{\code{level}}{a factor with levels \code{} \code{<1.0 mg/L} \code{<1.03 mmol/L} \code{<1.80 mmol/L} \code{<25 kg/m2} \code{<60 ml/min/1.73m2} \code{<65 years} \code{>=1.03-<1.55 mmol/L} \code{>=1.55 mmol/L} \code{>=1.80-<2.58 mmol/L} \code{>=2.58 mmol/L} \code{>=3} \code{>=30 kg/m2} \code{>=60 ml/min/1.73m2} \code{>=75 years} \code{>3.0 mg/L} \code{1} \code{1.0-3.0 mg/L} \code{2} \code{25-<30 kg/m2} \code{65-74 years} \code{American Indian or Alaska Native} \code{Asia/Pacific} \code{Asian} \code{Black} \code{Central/South/South East Asian} \code{Current smoker} \code{East Asian/Japanese} \code{Eastern Europe} \code{Female} \code{First tertile} \code{Former smoker} \code{High} \code{Hispanic or Latino} \code{Level 1: Males: <=0.90, Females: <=0.83} \code{Level 2: Males: >0.90-<=0.95, Females: >0.83-<=0.9} \code{Level 3: Males: >0.95, Females: >0.90} \code{Male} \code{Native Hawaiian or Other Pacific Islander} \code{Never smoked} \code{No} \code{Non-US} \code{Non-white} \code{North America} \code{Not Hispanic or Latino} \code{Other} \code{Other Mixed Race} \code{Recent} \code{Remote} \code{Second tertile} \code{South America} \code{Target} \code{Third tertile} \code{US} \code{Western Europe} \code{White} \code{Yes}}
    \item{\code{subgroup}}{a factor with levels \code{Additional pred. of CV risk: Age>=60 years} \code{Additional pred. of CV risk: Current or previous smoker} \code{Additional pred. of CV risk: Diabetes req. pharm.} \code{Additional pred. of CV risk: HDL-C <40 mg/dL} \code{Additional pred. of CV risk: Polyvascular disease} \code{Additional pred. of CV risk: Sig. renal dysfunction} \code{Age at randomization} \code{Baseline ACE inhibitor or Angiotensin II receptor blocker use} \code{Baseline aspirin use} \code{Baseline beta-blocker use} \code{Baseline blood pressure} \code{Baseline diabetes status} \code{Baseline eGFR} \code{Baseline HDL-C} \code{Baseline hs C-reactive protein} \code{Baseline LDL-C} \code{Baseline Lp-PLA2 Activity} \code{Baseline P2Y12 inhibitor use} \code{Baseline statin use} \code{Body mass index at baseline} \code{Ethnicity} \code{Family history of CHD} \code{FDA race group} \code{Gender} \code{No. Additional pred. of CV risk} \code{Qual. diag.: Multivessel CHD} \code{Qual. diag.: Prior Coronary Revascularization} \code{Qual. diag.: Prior MI} \code{Race group} \code{Race group collapsed} \code{Region} \code{Smoking status} \code{Time from CHD event to randomization} \code{United States Subgroup} \code{Waist/hip ratio at baseline}}
  }
}
\details{
No details.
}


\examples{
data(forest.data)
## maybe str(forest.data) ; plot(forest.data) ...
}
\keyword{datasets}
