% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{forest.plot}
\alias{forest.plot}
\title{forest.plot}
\usage{
forest.plot(parent.df = working.df, y.rank.col = "rank", Point.Est = "hr",
  Lower.CI = "low", Upper.CI = "high", y.label.rank.col = "rank",
  y.label.col = "subcategory", x.label = "Estimate", y.label = "Item",
  log.trans = TRUE, x.limits = c(0.21, 5), x.ticks = 2^(-2:2),
  y.limits = NULL, category.col = "category",
  background.palette = c("red", "blue"), category.palette = c("red",
  "blue"), shape.palette = c(16, 16), flip.palette = FALSE)
}
\arguments{
\item{y.rank.col}{column holding ranks for line items in forest/dot/table plots}

\item{Point.Est}{= "Point_Est",}

\item{Lower.CI}{column holding lower limit of CI}

\item{Upper.CI}{column holding upper limit of CI}

\item{y.label.rank.col}{column holding ranks for labels in forest/dot/table plots}

\item{y.label.col}{column holding labels for forest/dot/table plots}

\item{x.label}{value gets passed to labs}

\item{y.label}{value gets passed to labs}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{y.limits}{passed to scale_y_continuous}

\item{category.col}{parent.df column associated with explantory variable}

\item{background.palette}{palette gets passed to scale_fill_manual}

\item{category.palette}{colors assoicated with categorical variable}

\item{shape.palette}{values passed to scale_shape_manual}

\item{flip.palette}{logical; if TRUE it reverse the order of colors used for background}

\item{logtrans}{Logical; if true log transformation is applied to x axis (ensure x.limits are positive!)}
}
\description{
A function for plotting forest plots offering compatiability with table.plot and dot.plot.
}
\author{
Greg Cicconetti
}

