% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{gcurve}
\alias{gcurve}
\title{gcurve}
\usage{
gcurve(expr, from = NULL, to = NULL, n = 101, add = FALSE, type = "l",
  xname = "x", xlab = xname, ylab = NULL, log = NULL, xlim = NULL,
  category = NULL, ...)
}
\arguments{
\item{expr}{The name of a function, or a \link{call} or an
    \link{expression} written as a function of \code{x} which will
    evaluate to an object of the same length as \code{x}.}

\item{from}{the range over which the function will be plotted.}

\item{n}{integer; the number of x values at which to evaluate.}

\item{add}{logical; if \code{TRUE} add to an already existing plot; if
    \code{NA} start a new plot taking the defaults for the limits and
    log-scaling of the x-axis from the previous plot.  Taken as
    \code{FALSE} (with a warning if a different value is supplied) if no
    graphics device is open.}

\item{type}{plot type: see \code{\link{plot.default}}.}

\item{xname}{character string giving the name to be used for the x axis.}

\item{xlab}{labels and \link{graphical parameters}
    can also be specified as arguments.  See \sQuote{Details} for the
    interpretation of the default for \code{log}.

    For the \code{"function"} method of \code{plot}, \code{\dots} can
    include any of the other arguments of \code{curve}, except
    \code{expr}.
  }

\item{xlim}{\code{NULL} or a numeric vector of length 2;
    if non-\code{NULL} it provides the defaults for \code{c(from, to)}
    and, unless \code{add = TRUE}, selects the x-limits of the plot -- see
    \code{\link{plot.window}}.}

\item{category}{option to add a column populated with a factor}
}
\description{
A function to exploit base R's curve function.  This returns a data.frame holding x and y values returned from a call to curve, but suppress the plotting of that function
}
\examples{
\dontrun{
curve(dnorm(x, mean=0, sd=1), from=-4, to = 4, n= 1001)
ggplot(gcurve(expr = dnorm(x, mean=0, sd=1),from=-4, to = 4, n= 1001,
category="Standard Normal"), aes(x=x, y=y)) + geom_line()
}
}
\author{
Greg Cicconetti
}

