% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/builders.R
\name{get.top.xaxis}
\alias{get.top.xaxis}
\title{get.top.xaxis}
\usage{
get.top.xaxis(p1, p2)
}
\arguments{
\item{p1}{ggplot object with bottom x-axis}

\item{p2}{ggplot object with intended top x-axis in bottom position}
}
\description{
This takes two ggplot objects, steals the bottom x-axis from 2nd object and returns a gtable object with that bottom x-axis per object 1 and top x-axis per object 2
}

