% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{graphic.params}
\alias{graphic.params}
\title{Standard graphics names}
\usage{
graphic.params(at.risk.palette, background.palette, category.col,
  category.label, category.palette, censor.col, centime.col, interior,
  interior.h, interior.w, label.col, line.col, line.size, linetype.palette,
  ncol, nrow, page.height, page.width, parent.df, right.margin, left.margin,
  top.margin, bottom.margin, shape.label, shape.palette, test.dim, x.label,
  x.limits, x.ticks, x.ticks.labels, y.col, y.label, y.limits, ymin.col,
  ymax.col, y.ticks)
}
\arguments{
\item{at.risk.palette}{colors to be assocated with categorical variable in accompanying at.risk table}

\item{background.palette}{palette gets passed to scale_fill_manual}

\item{category.col}{parent.df column associated with explantory variable}

\item{category.label}{passed to x-axis label}

\item{category.palette}{colors assoicated with categorical variable}

\item{censor.col}{name of parent.df column associated with censor variable}

\item{centime.col}{name of parent.df column associated with censored time}

\item{interior}{a list of nrow*ncol grobs/ggplot objects to be displayed in the grid, ordered by row then col}

\item{interior.h}{a vector summing to 1 to indicate how to partition the heights}

\item{interior.w}{a vector summing to 1 to indicate how to partition the widths}

\item{label.col}{name of parent.df column containing labels}

\item{line.col}{name of parent.df column associated with linetype}

\item{line.size}{value gets passed to size within geom_line, geom_step}

\item{linetype.palette}{values passed to scale_linetype_manual}

\item{ncol}{number of columns for the grid of graphics being built by build.page}

\item{nrow}{number of rows for the grid of graphics being built by build.page}

\item{page.height}{used by build.page and annotate.page}

\item{page.width}{used by build.page and annotate.page}

\item{parent.df}{data.frame used by ggplot}

\item{right.margin}{used by build.page and annotate.page}

\item{left.margin}{used by build.page and annotate.page}

\item{top.margin}{used by build.page and annotate.page}

\item{bottom.margin}{used by build.page and annotate.page}

\item{shape.label}{value sets passed to labs}

\item{shape.palette}{values passed to scale_shape_manual}

\item{test.dim}{logical. Assists with figure development. If TRUE it makes a call to grid.show.layout.}

\item{x.label}{value gets passed to labs}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{x.ticks.labels}{passed to scale_x_continuous}

\item{y.col}{parent.df column associated with response vairable}

\item{y.label}{value gets passed to labs}

\item{y.limits}{passed to scale_y_continuous}

\item{ymin.col}{name of parent.df column assoicated with ymin}

\item{ymax.col}{name of parent.df column associated with ymax}

\item{y.ticks}{passed to scale_y_continuous}

\item{text.col1}{name of column holding text for column 1}

\item{text.col2}{name of column holding text for column 2; can be NULL}

\item{text.col3}{name of column holding text for column 3; can be NULL}

\item{text.col4}{name of column holding text for column 4; can be NULL}

\item{text.size}{value gets passed to geom_text}

\item{y.digits}{passed to scale_y_continuous label's, fmt}

\item{y.label.rank.col}{column holding ranks for labels in forest/dot/table plots}

\item{y.label.col}{column holding labels for forest/dot/table plots}

\item{y.rank.col}{column holding ranks for line items in forest/dot/table plots}
}
\description{
This is a dummy function, purpose is to serve as repositiory for function parameter names.
}

