% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{graphic.params}
\alias{graphic.params}
\title{Standard graphics names}
\usage{
graphic.params(at.risk.palette, background.palette, category.col,
  category.label, category.palette, censor.col, centime.col, label.col,
  line.col, line.size, linetype.palette, page.height, page.width, right.margin,
  left.margin, top.margin, bottom.margin, shape.label, shape.palette, x.label,
  x.limits, x.ticks, x.ticks.labels, y.col, y.label, y.limits, ymin.col,
  ymax.col, y.ticks)
}
\arguments{
\item{at.risk.palette}{colors to be assocated with categorical variable in accompanying at.risk table}

\item{background.palette}{palette gets passed to scale_fill_manual}

\item{category.col}{parent.df column associated with explantory variable}

\item{category.label}{passed to x-axis label}

\item{category.palette}{colors assoicated with categorical variable}

\item{censor.col}{name of parent.df column associated with censor variable}

\item{centime.col}{name of parent.df column associated with censored time}

\item{label.col}{name of parent.df column containing labels}

\item{line.col}{name of parent.df column associated with linetype}

\item{line.size}{value gets passed to size within geom_line, geom_step}

\item{linetype.palette}{values passed to scale_linetype_manual}

\item{page.height}{used by build.page and annotate.page}

\item{page.width}{used by build.page and annotate.page}

\item{right.margin}{used by build.page and annotate.page}

\item{left.margin}{used by build.page and annotate.page}

\item{top.margin}{used by build.page and annotate.page}

\item{bottom.margin}{used by build.page and annotate.page}

\item{shape.label}{value sets passed to labs}

\item{shape.palette}{values passed to scale_shape_manual}

\item{x.label}{value gets passed to labs}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{x.ticks.labels}{passed to scale_x_continuous}

\item{y.col}{parent.df column associated with response vairable}

\item{y.label}{value gets passed to labs}

\item{y.limits}{passed to scale_y_continuous}

\item{ymin.col}{name of parent.df column assoicated with ymin}

\item{ymax.col}{name of parent.df column associated with ymax}

\item{y.ticks}{passed to scale_y_continuous}

\item{text.size}{value gets passed to geom_text}

\item{y.digits}{passed to scale_y_continuous label's, fmt}

\item{y.label.rank.col}{column holding ranks for labels in forest/dot/table plots}

\item{y.label.col}{column holding labels for forest/dot/table plots}

\item{y.rank.col}{column holding ranks for line items in forest/dot/table plots}
}
\description{
This is a dummy function, purpose is to serve as repositiory for function parameter names.
}

