% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{km.plot}
\alias{km.plot}
\title{km.plot}
\usage{
km.plot(parent.df, censor.col = "CENSOR", centime.col = "CENTIME",
  category.col = "REGRAP", category.palette = rainbow(5),
  at.risk.palette = rainbow(5), category.label = "Treatment Group",
  nsubj.plot.label = "Number at Risk", linetype.palette = 1:6,
  x.label = "Time Since Randomization", y.label = "Percetage of Subjects",
  x.limits = c(0, 48), x.ticks = seq(0, 48, 3), y.ticks = seq(0, 1, 0.2),
  y.limits = c(0, 1), line.size = 0.75, fromthetop = FALSE,
  text.size = 4)
}
\arguments{
\item{censor.col}{name of parent.df column associated with censor variable}

\item{centime.col}{name of parent.df column associated with censored time}

\item{category.col}{parent.df column associated with explantory variable}

\item{category.palette}{colors assoicated with categorical variable}

\item{at.risk.palette}{colors to be assocated with categorical variable in accompanying at.risk table}

\item{category.label}{passed to x-axis label}

\item{linetype.palette}{values passed to scale_linetype_manual}

\item{x.label}{value gets passed to labs}

\item{y.label}{value gets passed to labs}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{y.ticks}{passed to scale_y_continuous}

\item{y.limits}{passed to scale_y_continuous}

\item{line.size}{value gets passed to size within geom_line, geom_step}

\item{fromthetop}{logical.  If TRUE KM curve decends from 1, if FALSE KM curve ascends from 0. Ensure you have an appropriate censor.col passed above!}

\item{text.size}{value gets passed to geom_text}
}
\description{
A function to create km plots
}
\author{
Greg Cicconetti
}

