% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{nsubj.plot}
\alias{nsubj.plot}
\title{nsubj.plot}
\usage{
nsubj.plot(parent.df = working.df, category.palette = c("red", "blue"),
  x.label = "Number of Subjects", y.label = "Treatment\\nGroup",
  text.size = 4, x.col = "XVALUES", text.col = "N",
  category.col = "CATEGORY", x.limits = c(0.5, 18),
  x.ticks = unique(parent.df$XVALUES),
  x.ticks.labels = unique(parent.df$XVALUES))
}
\arguments{
\item{parent.df}{incoming data.frame with special structure}

\item{category.palette}{colors assoicated with categorical variable}

\item{x.label}{value gets passed to labs}

\item{y.label}{value gets passed to labs}

\item{text.size}{value gets passed to geom_text}

\item{category.col}{parent.df column associated with explantory variable}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{x.ticks}{value gets passed to scale_x_continuous}

\item{x.ticks.labels}{passed to scale_x_continuous}
}
\description{
A function to create tables to accompany KMs and lineplots
}
\examples{
\dontrun{
}
}
\author{
Greg Cicconetti/David Wade
}

