% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{refresh.outputplan}
\alias{refresh.outputplan}
\title{Refresh the Output Plan}
\usage{
refresh.outputplan(loadplan = TRUE,
  filename = "C:/Users/gqc70045/Documents/LOCALR/Darapladib Graphics/Stability Graphics/refdata/outputplan_stability 22NOV13.csv")
}
\arguments{
\item{loadplan}{logical; if true then it loads from the filename, else it presumes outputplan is an object in Global Environment}

\item{filename}{Should point to the outputplan_study.csv typically located in the /code directory.}
}
\description{
Reloads outputplan_study.csv file and applies canonical formatting changes.
}
\details{
Ensure all columns are read in as character vectors. Ensure all missing entries are replaced with blank character string. Ensure all escape characters for carrige returns are respected. Grabs the 'modified time' from file attributes associated with .csv files named in the outputplan.
}
\author{
Greg Cicconetti
}

