% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{start.session.log}
\alias{start.session.log}
\title{start.session.log}
\usage{
\method{start}{session.log}(outputfile = "example.PDF")
}
\arguments{
\item{outputfile}{passed to name the session history log file}
}
\description{
A function to start logging the session history for a graphic driver run
}
\details{
Note that the stop.session.log function is used to stop the logging and save the log file.
}
\section{Value}{

\describe{
No objects are returned by this function.
}
}
\author{
David Wade
}

