% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphics.r
\name{table.plot}
\alias{table.plot}
\title{table.plot}
\usage{
table.plot(parent.df = working.df, y.rank.col = "Subcategory",
  category.col = "Treatment", text.col1 = "Point_Est", text.col2 = "LCI",
  text.col3 = "UCI", text.col4 = NULL, text.size = 12,
  xtick.labs = c("Estimate", "LCI", "UCI"), x.limits = NULL,
  y.limits = NULL, x.label = "Text", y.label = "Item",
  y.label.rank.col = "rank", y.label.col = "subcategory",
  category.palette = c("red", "blue"))
}
\arguments{
\item{y.rank.col}{column holding ranks for line items in forest/dot/table plots}

\item{category.col}{parent.df column associated with explantory variable}

\item{text.col1}{name of column holding text for column 1}

\item{text.col2}{name of column holding text for column 2; can be NULL}

\item{text.col3}{name of column holding text for column 3; can be NULL}

\item{text.col4}{name of column holding text for column 4; can be NULL}

\item{text.size}{value gets passed to geom_text}

\item{x.limits}{value gets passed to scale_x_continuous}

\item{y.limits}{passed to scale_y_continuous}

\item{x.label}{value gets passed to labs}

\item{y.label}{value gets passed to labs}

\item{y.label.rank.col}{column holding ranks for labels in forest/dot/table plots}

\item{y.label.col}{column holding labels for forest/dot/table plots}

\item{category.palette}{colors assoicated with categorical variable}
}
\description{
A function for plotting columns of text in a figure offering compatiability with forest.plot and dot.plot.
}
\author{
Greg Cicconetti
}

